Fix layout calculation
git-svn-id: https://svn.code.sf.net/p/ourorgan/svn/trunk@1857 69509848-2a73-49d5-a893-c4d8c934025a

+++ src/grandorgue/GOGUILayoutEngine.cpp
@@ -105,7 +105,7 @@ int GOGUILayoutEngine::GetCenterWidth()
 
 int GOGUILayoutEngine::GetCenterX()
 {
-	return (m_metrics.GetScreenWidth() - GetCenterWidth()) >> 1;
+	return ((int)m_metrics.GetScreenWidth() - GetCenterWidth()) >> 1;
 }
 
 int GOGUILayoutEngine::GetCenterY()
